requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0
sentencepiece>=0.1.99
sacremoses>=0.0.53

# Hugging Face ecosystem
accelerate>=0.20.0
evaluate>=0.4.0
peft>=0.4.0

# Indian language specific libraries
indic-nlp-library>=0.81
polyglot>=16.7.4
indicnlp>=0.0.1

# Data processing and utilities
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Text processing
nltk>=3.8
spacy>=3.6.0
regex>=2023.6.3
ftfy>=6.1.1
langdetect>=1.0.9

# Web scraping and data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
scrapy>=2.9.0
selenium>=4.10.0

# Configuration and logging
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0
wandb>=0.15.0
tensorboard>=2.13.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.4.0

# Jupyter and notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=6.5.0

# Storage and databases
sqlalchemy>=2.0.0
pymongo>=4.4.0

# Performance and optimization
numba>=0.57.0
joblib>=1.3.0

# Visualization
wordcloud>=1.9.0
networkx>=3.1

# Optional: GPU support (uncomment if using CUDA)
# torch-audio>=2.0.0
# torchvision>=0.15.0
